+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add hyperlinks to citations when `hyperref` is loaded.
 - Add backref.
+- Add journal abbreviation.
 
 ### Fixed